Create 0068. Text Justification
Written in Java.

https://leetcode.com/problems/text-justification/

"Hardest" one accomplished, but not very elegant. Brute forced in creating list with a bunch of redundancies. Could probably be done without so many arraylists. Some of the variables should be renamed, and comments need to be added to all of it.

Cheats a little on the last one, but handles the edge case where the final line is exactly the right width to end on the string.