Create 0030. Substring with Concatenation of All Words
Initial version was similar, but brute force. This is an improved version that breaks each string into a map, then compares it to the "word" map. The map also builds in amounts, for when words are repeated, which prevents edge cases with large numbers of the same letter from eating up too much time in comparing.

https://leetcode.com/problems/substring-with-concatenation-of-all-words/